Convert Long-Format Data into Wide-Format for Reporting (Internal)

Converts a long-format data frame into a wide-format structure, applying optional transformations such as calculating totals, renaming columns, and rounding numeric values. It ensures grouping columns are preserved while transforming the data into a structured table for reporting.

Usage

.process_detail_data(
  df,
  wide_col,
  group_cols,
  rename_mapping,
  total_column,
  decimal
)

Arguments

df

A data frame to be transformed.

wide_col

Character. The column whose unique values will become new column headers in the wide-format table.

group_cols

Character vector. Column(s) to retain as row identifiers in the wide-format structure.

rename_mapping

Named list. Specifies mapping for renaming grouping columns, where names are existing column names, and values are new column names.

total_column

Logical. If TRUE, adds a "Total" column that sums all numeric columns. Default is FALSE.

decimal

Numeric. Number of decimal places to round numeric columns. Default is 2.

Value

A wide-format data frame with transformed column headers, optional total values, renamed columns, and rounded numeric values.

Details

This function preserves important grouping columns while converting a data frame into wide format. It also ensures numeric columns are properly rounded and formatted, and allows renaming of specific grouping variables for clearer reporting.

If total_column = TRUE, the function calculates the total of all numeric columns and adds a "Total" column.

If rename_mapping is provided, the function renames matching columns.

Sorting is applied to ensure proper column arrangement.

Author

Pattawee Puangchit