Print Plot and Export Configuration Snippets

Retrieve full configuration code as a list for applying in the plot styling and export settings.

Usage

get_all_config(
  plot_style = "default",
  plot_config = TRUE,
  export_config = TRUE
)

Arguments

plot_style

Character. Plot style to use (currently only `"default"` is supported).

plot_config

Logical. If `TRUE`, prints the plot style configuration.

export_config

Logical. If `TRUE`, prints the export configuration.

Value

List of configuration snippets to the console.

Details

Onece printing into the console, users can simply copy and paste the entire list of configurations, rename it (if needed), and use it in your plot functions directly.

Author

Pattawee Puangchit

Examples

# Input Path:
input_path <- system.file("extdata/in", package = "GTAPViz")
sl4.plot.data <- readRDS(file.path(input_path, "sl4.plot.data.rds"))
#> Warning: cannot open compressed file 'C:/Users/b_pat/AppData/Local/R/win-library/4.4/GTAPViz/extdata/in/sl4.plot.data.rds', probable reason 'No such file or directory'
#> Error in gzfile(file, "rb"): cannot open the connection

# Retrive configurations
get_all_config()
#> $plot_style_config
#>                Topic                      Arguments
#> 1              Title                     show_title
#> 2                                        title_face
#> 3                                        title_size
#> 4                                       title_hjust
#> 5                                 add_unit_to_title
#> 6                                      title_margin
#> 7                                      title_format
#> 8             X-Axis              show_x_axis_title
#> 9                                 x_axis_title_face
#> 10                                x_axis_title_size
#> 11                              x_axis_title_margin
#> 12                               show_x_axis_labels
#> 13                                 x_axis_text_face
#> 14                                 x_axis_text_size
#> 15                                x_axis_text_angle
#> 16                                x_axis_text_hjust
#> 17                               x_axis_description
#> 18            Y-Axis              show_y_axis_title
#> 19                                y_axis_title_face
#> 20                                y_axis_title_size
#> 21                              y_axis_title_margin
#> 22                               show_y_axis_labels
#> 23                                 y_axis_text_face
#> 24                                 y_axis_text_size
#> 25                                y_axis_text_angle
#> 26                                y_axis_text_hjust
#> 27                               y_axis_description
#> 28                   show_axis_titles_on_all_facets
#> 29      Value Labels              show_value_labels
#> 30                                 value_label_face
#> 31                                 value_label_size
#> 32                             value_label_position
#> 33                       value_label_decimal_places
#> 34            Legend                    show_legend
#> 35                                show_legend_title
#> 36                                  legend_position
#> 37                                legend_title_face
#> 38                                 legend_text_face
#> 39                                 legend_text_size
#> 40       Panel Strip                     strip_face
#> 41                                  strip_text_size
#> 42                                 strip_background
#> 43                                strip_text_margin
#> 44      Panel Layout                  panel_spacing
#> 45                                       panel_rows
#> 46                                       panel_cols
#> 47                                            theme
#> 48            Colors                     color_tone
#> 49                               color_palette_type
#> 50                                   positive_color
#> 51                                   negative_color
#> 52                                 background_color
#> 53                                       grid_color
#> 54                                show_grid_major_x
#> 55                                show_grid_major_y
#> 56                                show_grid_minor_x
#> 57                                show_grid_minor_y
#> 58         Zero Line                 show_zero_line
#> 59                                   zero_line_type
#> 60                                  zero_line_color
#> 61                                   zero_line_size
#> 62                               zero_line_position
#> 63         Bar Chart                      bar_width
#> 64                                      bar_spacing
#> 65    Scale Settings                    scale_limit
#> 66                                  scale_increment
#> 67   Scale Expansion               expansion_y_mult
#> 68                                 expansion_x_mult
#> 69 Font Size Control                  all_font_size
#> 70      Data Sorting             sort_data_by_value
#> 71       Plot Margin                    plot.margin
#>                                                                               Default.Value
#> 1                                                                                      TRUE
#> 2                                                                                      bold
#> 3                                                                                        20
#> 4                                                                                       0.5
#> 5                                                                                      TRUE
#> 6                                                                           c(10, 0, 10, 0)
#> 7  list(type="standard" [options: standard/prefix/suffix/full/dynamic], text="", sep=" - ")
#> 8                                                                                      TRUE
#> 9                                                                                      bold
#> 10                                                                                       16
#> 11                                                                          c(25, 25, 0, 0)
#> 12                                                                                     TRUE
#> 13                                                                                    plain
#> 14                                                                                       14
#> 15                                                                                        0
#> 16                                                                                        0
#> 17                                                                                         
#> 18                                                                                     TRUE
#> 19                                                                                     bold
#> 20                                                                                       16
#> 21                                                                          c(25, 25, 0, 0)
#> 22                                                                                     TRUE
#> 23                                                                                    plain
#> 24                                                                                       14
#> 25                                                                                        0
#> 26                                                                                        0
#> 27                                                                                         
#> 28                                                                                     TRUE
#> 29                                                                                     TRUE
#> 30                                                                                    plain
#> 31                                                                                        5
#> 32                                                                                    above
#> 33                                                                                        2
#> 34                                                                                    FALSE
#> 35                                                                                    FALSE
#> 36                                                                                   bottom
#> 37                                                                                     bold
#> 38                                                                                    plain
#> 39                                                                                       14
#> 40                                                                                     bold
#> 41                                                                                       16
#> 42                                                                                lightgrey
#> 43                                                                          c(10, 0, 10, 0)
#> 44                                                                                        2
#> 45                                                                                     NULL
#> 46                                                                                     NULL
#> 47                                                                                     NULL
#> 48                                                                                     NULL
#> 49                                                                            "qualitative"
#> 50                                                                                  #2E8B57
#> 51                                                                                  #CD5C5C
#> 52                                                                                    white
#> 53                                                                                   grey90
#> 54                                                                                    FALSE
#> 55                                                                                    FALSE
#> 56                                                                                    FALSE
#> 57                                                                                    FALSE
#> 58                                                                                     TRUE
#> 59                                                                                   dashed
#> 60                                                                                    black
#> 61                                                                                      0.5
#> 62                                                                                        0
#> 63                                                                                      0.9
#> 64                                                                                      0.9
#> 65                                                                                     NULL
#> 66                                                                                     NULL
#> 67                                                                             c(0.05, 0.1)
#> 68                                                                            c(0.05, 0.05)
#> 69                                                                                        1
#> 70                                                                                    FALSE
#> 71                                                                        c(10, 25, 10, 10)
#>       Input.Format
#> 1          logical
#> 2        character
#> 3          numeric
#> 4          numeric
#> 5          logical
#> 6   numeric vector
#> 7             list
#> 8          logical
#> 9        character
#> 10         numeric
#> 11  numeric vector
#> 12         logical
#> 13       character
#> 14         numeric
#> 15         numeric
#> 16         numeric
#> 17       character
#> 18         logical
#> 19       character
#> 20         numeric
#> 21  numeric vector
#> 22         logical
#> 23       character
#> 24         numeric
#> 25         numeric
#> 26         numeric
#> 27       character
#> 28         logical
#> 29         logical
#> 30       character
#> 31         numeric
#> 32       character
#> 33         numeric
#> 34         logical
#> 35         logical
#> 36       character
#> 37       character
#> 38       character
#> 39         numeric
#> 40       character
#> 41         numeric
#> 42       character
#> 43  numeric vector
#> 44         numeric
#> 45 NULL or numeric
#> 46 NULL or numeric
#> 47   ggplot2 theme
#> 48       character
#> 49       character
#> 50       character
#> 51       character
#> 52       character
#> 53       character
#> 54         logical
#> 55         logical
#> 56         logical
#> 57         logical
#> 58         logical
#> 59       character
#> 60         numeric
#> 61         numeric
#> 62         numeric
#> 63         numeric
#> 64         numeric
#> 65  numeric vector
#> 66         numeric
#> 67  numeric vector
#> 68  numeric vector
#> 69         numeric
#> 70         logical
#> 71  numeric vector
#>                                                                                                                                                                                                             Description
#> 1                                                                                                                                                                                          Show or hide the plot title.
#> 2                                                                                                                                                                      Font face for title ('bold', 'plain', 'italic').
#> 3                                                                                                                                                                                                   Font size of title.
#> 4                                                                                                                                                              Horizontal justification of title (0 = left, 1 = right).
#> 5                                                                                                                                                                                        Add unit information to title.
#> 6                                                                                                                                                                       Margin around title (top, right, bottom, left).
#> 7  List with components: 'type' (options: 'standard', 'prefix', 'suffix', 'full', 'dynamic'), 'text' (content to display or column names for dynamic titles), and 'sep' (separator for dynamic titles, default: ' - ').
#> 8                                                                                                                                                                                        Show or hide the x-axis title.
#> 9                                                                                                                                                                                           Font face for x-axis title.
#> 10                                                                                                                                                                                           Font size of x-axis title.
#> 11                                                                                                                                                               Margin around x-axis title (top, right, bottom, left).
#> 12                                                                                                                                                                                     Show or hide x-axis tick labels.
#> 13                                                                                                                                                                                    Font face for x-axis tick labels.
#> 14                                                                                                                                                                                     Font size of x-axis tick labels.
#> 15                                                                                                                                                                              Angle of x-axis tick labels in degrees.
#> 16                                                                                                                                                                      Horizontal justification of x-axis tick labels.
#> 17                                                                                                                                                                                     Optional description for x-axis.
#> 18                                                                                                                                                                                       Show or hide the y-axis title.
#> 19                                                                                                                                                                                          Font face for y-axis title.
#> 20                                                                                                                                                                                           Font size of y-axis title.
#> 21                                                                                                                                                               Margin around y-axis title (top, right, bottom, left).
#> 22                                                                                                                                                                                     Show or hide y-axis tick labels.
#> 23                                                                                                                                                                                    Font face for y-axis tick labels.
#> 24                                                                                                                                                                                     Font size of y-axis tick labels.
#> 25                                                                                                                                                                              Angle of y-axis tick labels in degrees.
#> 26                                                                                                                                                                      Horizontal justification of y-axis tick labels.
#> 27                                                                                                                                                                                     Optional description for y-axis.
#> 28                                                                                                                                                                                      Show axis titles on all facets.
#> 29                                                                                                                                                                                           Show or hide value labels.
#> 30                                                                                                                                                                                          Font face for value labels.
#> 31                                                                                                                                                                                           Font size of value labels.
#> 32                                                                                                                                                                Position of value labels ('above', 'outside', 'top').
#> 33                                                                                                                                                                            Number of decimal places in value labels.
#> 34                                                                                                                                                                                             Show or hide the legend.
#> 35                                                                                                                                                                                       Show or hide the legend title.
#> 36                                                                                                                                                            Position of the legend ('none', 'right', 'bottom', etc.).
#> 37                                                                                                                                                                                          Font face for legend title.
#> 38                                                                                                                                                                                           Font face for legend text.
#> 39                                                                                                                                                                                            Font size of legend text.
#> 40                                                                                                                                                                                    Font face for panel strip labels.
#> 41                                                                                                                                                                                     Font size of panel strip labels.
#> 42                                                                                                                                                                                    Background color of panel strips.
#> 43                                                                                                                                                         Margin around panel strip labels (top, right, bottom, left).
#> 44                                                                                                                                                                               Spacing between panels in centimeters.
#> 45                                                                                                                                                                                      Number of rows in panel layout.
#> 46                                                                                                                                                                                   Number of columns in panel layout.
#> 47                                                                                                                                                                                               Custom theme to apply.
#> 48                                                                                                                                                           Base color tone for the plot (e.g., 'academic', 'purdue').
#> 49                                                                                                                                                 Type of color palette ('qualitative', 'sequential', or 'diverging').
#> 50                                                                                                                                                                                           Color for positive values.
#> 51                                                                                                                                                                                           Color for negative values.
#> 52                                                                                                                                                                                        Background color of the plot.
#> 53                                                                                                                                                                                                 Color of grid lines.
#> 54                                                                                                                                                                                     Show major grid lines on x-axis.
#> 55                                                                                                                                                                                     Show major grid lines on y-axis.
#> 56                                                                                                                                                                                     Show minor grid lines on x-axis.
#> 57                                                                                                                                                                                     Show minor grid lines on y-axis.
#> 58                                                                                                                                                                                          Show or hide the zero line.
#> 59                                                                                                                                                               Line type for zero line ('solid', 'dashed', 'dotted').
#> 60                                                                                                                                                                                                  Color of zero line.
#> 61                                                                                                                                                                                         Line thickness of zero line.
#> 62                                                                                                                                                                                           Position of the zero line.
#> 63                                                                                                                                                                                                 Width of bars (0-1).
#> 64                                                                                                                                                                                      Spacing between groups of bars.
#> 65                                                                                                                                                                             Manual limits for value axis (min, max).
#> 66                                                                                                                                                                                       Step size for axis tick marks.
#> 67                                                                                                                                                                                     Expansion multiplier for y-axis.
#> 68                                                                                                                                                                                     Expansion multiplier for x-axis.
#> 69                                                                                                                     Master control for all font sizes. Values > 1 increase all fonts, values < 1 decrease all fonts.
#> 70                                                                                                                                                              Whether to sort data by value for better visualization.
#> 71                                                                                                                                                           Margins around the entire plot (top, right, bottom, left).
#>                                                                                                                   Example
#> 1                                                                                                       show_title = TRUE
#> 2                                                                                                     title_face = "bold"
#> 3                                                                                                         title_size = 20
#> 4                                                                                                       title_hjust = 0.5
#> 5                                                                                                add_unit_to_title = TRUE
#> 6                                                                                          title_margin = c(10, 0, 10, 0)
#> 7  title_format = list(\n  type = "standard", # options: standard/prefix/suffix/full/dynamic\n  text = "",\n  sep = ""\n)
#> 8                                                                                                show_x_axis_title = TRUE
#> 9                                                                                              x_axis_title_face = "bold"
#> 10                                                                                                 x_axis_title_size = 16
#> 11                                                                                  x_axis_title_margin = c(25, 25, 0, 0)
#> 12                                                                                              show_x_axis_labels = TRUE
#> 13                                                                                             x_axis_text_face = "plain"
#> 14                                                                                                  x_axis_text_size = 14
#> 15                                                                                                  x_axis_text_angle = 0
#> 16                                                                                                  x_axis_text_hjust = 0
#> 17                                                                                                x_axis_description = ""
#> 18                                                                                               show_y_axis_title = TRUE
#> 19                                                                                             y_axis_title_face = "bold"
#> 20                                                                                                 y_axis_title_size = 16
#> 21                                                                                  y_axis_title_margin = c(25, 25, 0, 0)
#> 22                                                                                              show_y_axis_labels = TRUE
#> 23                                                                                             y_axis_text_face = "plain"
#> 24                                                                                                  y_axis_text_size = 14
#> 25                                                                                                  y_axis_text_angle = 0
#> 26                                                                                                  y_axis_text_hjust = 0
#> 27                                                                                                y_axis_description = ""
#> 28                                                                                  show_axis_titles_on_all_facets = TRUE
#> 29                                                                                               show_value_labels = TRUE
#> 30                                                                                             value_label_face = "plain"
#> 31                                                                                                   value_label_size = 5
#> 32                                                                                         value_label_position = "above"
#> 33                                                                                         value_label_decimal_places = 2
#> 34                                                                                                    show_legend = FALSE
#> 35                                                                                              show_legend_title = FALSE
#> 36                                                                                             legend_position = "bottom"
#> 37                                                                                             legend_title_face = "bold"
#> 38                                                                                             legend_text_face = "plain"
#> 39                                                                                                  legend_text_size = 14
#> 40                                                                                                    strip_face = "bold"
#> 41                                                                                                   strip_text_size = 16
#> 42                                                                                         strip_background = "lightgrey"
#> 43                                                                                    strip_text_margin = c(10, 0, 10, 0)
#> 44                                                                                                      panel_spacing = 2
#> 45                                                                                                      panel_rows = NULL
#> 46                                                                                                      panel_cols = NULL
#> 47                                                                                                           theme = NULL
#> 48                                                                                                      color_tone = NULL
#> 49                                                                                     color_palette_type = "qualitative"
#> 50                                                                                             positive_color = "#2E8B57"
#> 51                                                                                             negative_color = "#CD5C5C"
#> 52                                                                                             background_color = "white"
#> 53                                                                                                  grid_color = "grey90"
#> 54                                                                                              show_grid_major_x = FALSE
#> 55                                                                                              show_grid_major_y = FALSE
#> 56                                                                                              show_grid_minor_x = FALSE
#> 57                                                                                              show_grid_minor_y = FALSE
#> 58                                                                                                  show_zero_line = TRUE
#> 59                                                                                              zero_line_type = "dashed"
#> 60                                                                                              zero_line_color = "black"
#> 61                                                                                                   zero_line_size = 0.5
#> 62                                                                                                 zero_line_position = 0
#> 63                                                                                                        bar_width = 0.9
#> 64                                                                                                      bar_spacing = 0.9
#> 65                                                                                                     scale_limit = NULL
#> 66                                                                                                 scale_increment = NULL
#> 67                                                                                        expansion_y_mult = c(0.05, 0.1)
#> 68                                                                                       expansion_x_mult = c(0.05, 0.05)
#> 69                                                                                                      all_font_size = 1
#> 70                                                                                              sort_data_by_value = TRUE
#> 71                                                                                        plot.margin = c(10, 25, 10, 10)
#> 
#> $export_config
#>           Topic Arguments Default.Value    Input.Format
#> 1 Export Config file_name  "gtap_plots"       character
#> 2                   width          NULL numeric or NULL
#> 3                  height          NULL numeric or NULL
#> 4                     dpi           300         numeric
#> 5                      bg       "white"       character
#> 6               limitsize         FALSE         logical
#>                                                   Description
#> 1      Base name for exported files. Default is 'gtap_plots'.
#> 2  Plot width in inches. Default is automatically calculated.
#> 3 Plot height in inches. Default is automatically calculated.
#> 4                  Resolution for PNG export. Default is 300.
#> 5                       Background color. Default is 'white'.
#> 6                    Whether to limit size. Default is FALSE.
#>                          Example
#> 1 file_name = "regional_impacts"
#> 2                     width = 12
#> 3                     height = 8
#> 4                      dpi = 600
#> 5                   bg = "white"
#> 6              limitsize = FALSE
#>