Generate Macro Plot with Flexible Title Options

Creates figures of macro variables from the Macros header in the SL4. Uses a separate dataframe to specify variables to plot with options for using either variable names, PlotTitle, or Description columns for plot titles.

Usage

macro_plot(
  data,
  plot_var = NULL,
  title_prefix = "",
  title_suffix = "",
  compare_by_experiment = FALSE,
  description_as_title = FALSE,
  output_dir = NULL,
  panel_rows = NULL,
  panel_cols = NULL,
  color_tone = NULL,
  separate_figure = FALSE,
  width = NULL,
  height = NULL,
  legend_position = "none"
)

Arguments

data

A data frame containing the full dataset with macroeconomic variables.

plot_var

A data frame with "Variable" column and optionally "PlotTitle" or "Description" columns specifying which variables to plot and their titles. If NULL, all variables in the data will be plotted with default titles.

title_prefix

Optional character string to prepend to the title.

title_suffix

Optional character string to append to the title.

compare_by_experiment

Logical. If TRUE, compares experiments within x-axis categories.

description_as_title

Logical. If TRUE, uses the "Description" column for plot titles instead of "PlotTitle". Default is FALSE.

output_dir

Optional character. Directory to save the output plot.

panel_rows

Optional numeric. Number of panel rows. If left blank, it will be automatically adjusted.

panel_cols

Optional numeric. Number of panel columns. If left blank, it will be automatically adjusted.

color_tone

Optional character. Base color for the plot. If left blank, a default color will be used.

separate_figure

Logical. If FALSE, combines multiple variables into one plot.

width

Optional numeric. Width of the output plot. If left blank, it will be automatically adjusted.

height

Optional numeric. Height of the output plot. If left blank, it will be automatically adjusted.

legend_position

Position of the legend: "none", "bottom", "top", "left", or "right"

Value

A ggplot object or a list of plots if separate_figure = TRUE.

Examples

if (FALSE) { # \dontrun{
macro_plot(Macros,
           plot_var = macro.map,
           compare_by_experiment = FALSE,
           color_tone = "grey",
           output_dir = output.folder,
           description_as_title = FALSE,
           panel_rows = 2,
           width = 45,
           separate_figure = FALSE,
           legend_position = "bottom")
} # }