Export Data as an Excel Pivot Table

Exports a dataset to an Excel file with both raw data and a generated pivot table.

Usage

pivot_table_with_filter(
  data,
  filter = NULL,
  rows = NULL,
  cols = NULL,
  data_fields = "Value",
  raw_sheet_name = "RawData",
  pivot_sheet_name = "PivotTable",
  dims = "A5",
  export = TRUE,
  output_path = getwd(),
  workbook_name = "GTAP_PivotTable.xlsx"
)

Arguments

data

Data frame. The dataset to be exported.

filter

Character vector (optional). Columns to be used as filter fields in the pivot table.

rows

Character vector (optional). Columns to be used as row fields in the pivot table.

cols

Character vector (optional). Columns to be used as column fields in the pivot table.

data_fields

Character. The data field(s) to be summarized in the pivot table (default: `"Value"`).

raw_sheet_name

Character. Name of the sheet containing raw data (default: `"RawData"`).

pivot_sheet_name

Character. Name of the sheet containing the pivot table (default: `"PivotTable"`).

dims

Character. Cell reference where the pivot table starts (default: `"A3"`).

export

Logical. Whether to save the Excel file (default: `TRUE`).

output_path

Character. Directory where the file should be saved (default: current working directory).

workbook_name

Character. Name of the output Excel file (default: `"GTAP_PivotTable.xlsx"`).

Value

An excel workbook object containing both raw data and the pivot table.

Details

This function creates an Excel workbook with: - A raw data sheet (`raw_sheet_name`) containing the provided dataset. - A pivot table sheet (`pivot_sheet_name`) generated based on specified row, column, and data fields.

If `export = TRUE`, the function saves the workbook to the specified `output_path`.

Author

Pattawee Puangchit

Examples

# \donttest{
# Input Path
input_path <- system.file("extdata/in", package = "GTAPViz")

# Prepares data from SL4 files for EXP1 and EXP2
auto_gtap_data(
  experiment = c("EXP1", "EXP2"),
  input_path = input_path,
  subtotal_level = FALSE,
  process_sl4_vars = c("qo", "qgdp"),
  process_har_vars = FALSE,
  mapping_info = "GTAPv7",
  plot_data = TRUE
)
#> All 2 requested experiment SL4 files are found.
#> Mapping method used: GTAPv7 
#> Processing SL4 Data
#> 
#> Summary of Processing:
#> SL4 Data processed successfully
#> GTAP data processing completed successfully!
#> 

# Generate comparison table
pivot_table_with_filter(
  data = sl4.plot.data[["2D"]][["Sector"]],
  filter = c("Variable", "Unit"),
  rows = c("Region", "Sector"),
  cols = c("Experiment"),
  data_fields = "Value",
  raw_sheet_name = "Raw_Data",
  pivot_sheet_name = "Sector_Pivot",
  export = TRUE,
  output_path = tempdir(),
  workbook_name = "Sectoral_Impact_Analysis.xlsx"
)
#> Excel file with pivot table exported to: /tmp/RtmpyxW0tf/Sectoral_Impact_Analysis.xlsx
#> A Workbook object.
#>  
#> Worksheets:
#>  Sheets: Raw_Data, Sector_Pivot 
#>  Write order: 1, 2
# }