Add Mapping Information

The description and unit are based on GTAPv7 model variables. Full list refer to https://www.gtap.agecon.purdue.edu/models/setsvariables.asp. If any variable is added apart from the default GTAPv7 model, the information will be left blank.

Usage

add_mapping_info(
  data_list,
  external_map = NULL,
  mapping = "GTAPv7",
  description_info = TRUE,
  unit_info = TRUE
)

Arguments

data_list

A list or data frame containing GTAP variables.

external_map

Optional. A data frame containing external mapping information with columns "Variable", "Description", and "Unit". If NULL, the default GTAPv7 mapping is used.

mapping

Optional. Either "GTAPv7", "No", "Yes", or "Mix". Determines the mapping approach. Default is "GTAPv7".

description_info

Optional. Logical. If TRUE, adds description information. Default is TRUE.

unit_info

Optional. Logical. If TRUE, adds unit information. Default is TRUE.

Value

A modified list or data frame with added description and/or unit information.

Details

For the "E1" header in HAR files, if using GTAPv7 or Mix mapping mode, it renames the "FORM" column to "Unit". When using "Yes" mapping mode with an external map, it will use provided Description and Unit values.

Author

Pattawee Puangchit

Examples

sl4_data1 <- HARplus::load_sl4x(system.file("extdata", "TAR10.sl4", package = "HARplus"))
result <- add_mapping_info(sl4_data1, mapping = "GTAPv7", unit_info = FALSE) # Add only Description
result <- add_mapping_info(sl4_data1, mapping = "GTAPv7", description_info = FALSE) # Add only Unit
custom_mapping <- data.frame(Variable = c("qgdp", "EV"),
                             Description = c("Real GDP", "Economic Value"),
                             Unit = c("Percent", "million USD"))
result <- add_mapping_info(sl4_data1, external_map = custom_mapping, mapping = "Mix") # Custom + GTAPv7