Generate Detailed Plot

Creates a detailed plot reporting all regions or sectors depending on the variable selection. Uses a separate dataframe to specify variables to plot and their titles.

detail_plot(
  data,
  plot_var = NULL,
  y_axis_from,
  figure_separate_by,
  title_prefix = "",
  title_suffix = "",
  panel_rows = NULL,
  panel_cols = NULL,
  positive_color = "#2E8B57",
  negative_color = "#CD5C5C",
  output_dir = NULL,
  top_impact = NULL,
  width = NULL,
  height = NULL
)

Arguments

data

A data frame containing the full dataset.

plot_var

A data frame with "Variable" and "PlotTitle" columns specifying which variables to plot and their titles. If NULL, all variables in the data will be plotted with default titles.

y_axis_from

Character vector. Column names to use for the y-axis.

figure_separate_by

Character. Column used to separate different plots.

title_prefix

Optional character string to prepend to the title.

title_suffix

Optional character string to append to the title.

panel_rows

Optional numeric. Number of panel rows. If left blank, it will be automatically adjusted.

panel_cols

Optional numeric. Number of panel columns. If left blank, it will be automatically adjusted.

positive_color

Character. Color for positive values.

negative_color

Character. Color for negative values.

output_dir

Optional character. Directory to save the output plot.

top_impact

Optional numeric. Number of top impact factors to include. If left blank, all factors will be included.

width

Optional numeric. Width of the output plot. If left blank, it will be automatically adjusted.

height

Optional numeric. Height of the output plot. If left blank, it will be automatically adjusted.

Value

A ggplot object or NULL if no plots could be generated.