Rename GTAP Bilateral Trade Columns

Renames bilateral trade columns in GTAP data to standardized names, ensuring consistency in regional trade flows.

Usage

rename_GTAP_bilateral(data)

Arguments

data

Data structure containing GTAP bilateral trade data.

Value

The same data structure with renamed bilateral trade columns.

See also

Author

Pattawee Puangchit

Examples

# Load Sample Data:
sl4_data1 <- HARplus::load_sl4x(system.file("extdata/in", "EXP1.sl4", 
                                package = "GTAPViz"))
# Get Data by Variable Name
sl4_data1 <- HARplus::get_data_by_var("qxs",sl4_data1)

# Rename bilateral trade columns in a GTAP dataset
gtap_data <- rename_GTAP_bilateral(sl4_data1)