Change Specific Values in a Dataframe Column

Replaces specified values in a given column with new values. Only values listed in the original_values parameter will be changed.

rename_value(data, column_name, original_values, new_values)

Arguments

data

A data frame containing the column to modify.

column_name

Character. The name of the column where values will be changed.

original_values

Character vector. The values to be replaced.

new_values

Character vector. The new values corresponding to original_values.

Value

A modified data frame with specified values replaced.

Author

Pattawee Puangchit

Examples

har_data <- HARplus::load_harx(c("globalcgds", "vgdpwld", "WEV"))
#> Error in file(con, "rb"): invalid 'description' argument
har_data <- HARplus::get_data_by_var(c("globalcgds", "vgdpwld", "WEV"), har_data)
#> Error: object 'har_data' not found
modified_data <- rename_value(har_data[["A"]], "Variable",
                               c("alloc_A1", "ENDWB1", "tech_C1"),
                               c("Allocation", "Endowment", "Technology"))
#> Error: object 'har_data' not found